+++ app/controllers/guild_invites_controller.rb
@@ -1,8 +1,15 @@
 class GuildInvitesController < ApplicationController
   before_action :find_guild, only: %i[new]
+  before_action :find_guild_invite, only: %i[destroy]
 
   def new
     authorize! @from_guild, with: GuildInvitePolicy, context: { guild: @invite_creator, character: @invite_creator }
+  end
+
+  def destroy
+    authorize! @guild_invite.from_guild.to_s, with: GuildInvitePolicy, context: { guild: @guild_invite.guild, character: @guild_invite.character }
+    @guild_invite.destroy
+    redirect_to management_guild_path(@guild_invite.guild.slug)
   end
 
   private
@@ -19,5 +26,10 @@ class GuildInvitesController < ApplicationController
     else
       render_error('Guild ID or Character ID must be presented')
     end
+  end
+
+  def find_guild_invite
+    @guild_invite = GuildInvite.find_by(id: params[:id])
+    render_error('Object is not found') if @guild_invite.nil?
   end
 end