+++ config/routes.rb
@@ -64,7 +64,7 @@ Rails.application.routes.draw do
       get :decline, on: :member
     end
     resources :deliveries, only: %i[new destroy]
-    resources :guild_invites, only: %i[new]
+    resources :guild_invites, only: %i[new destroy]
 
     root to: 'welcome#index'
   end