+++ spec/controllers/guild_invites_controller_spec.rb
@@ -94,4 +94,60 @@ RSpec.describe GuildInvitesController, type: :controller do
       get :new, params: { locale: 'en', deliveriable_id: guild.id, deliveriable_type: 'Guild' }
     end
   end
+
+  describe 'DELETE#destroy' do
+    let!(:guild) { create :guild }
+    let!(:character) { create :character }
+    let!(:guild_invite) { create :guild_invite, guild: guild, character: character, from_guild: true }
+
+    it_behaves_like 'User Auth'
+
+    context 'for logged user' do
+      sign_in_user
+      let!(:user_character) { create :character, user: @current_user, guild: guild }
+
+      context 'for unexisted guild invite' do
+        it 'renders error page' do
+          delete :destroy, params: { locale: 'en', id: 'unexisted' }
+
+          expect(response).to render_template 'shared/error'
+        end
+      end
+
+      context 'for existed guild invite' do
+        context 'for invalid access' do
+          let(:request) { delete :destroy, params: { locale: 'en', id: guild_invite.id } }
+
+          it 'does not delete guild invite' do
+            expect { request }.to_not change(GuildInvite, :count)
+          end
+
+          it 'and renders error page' do
+            request
+
+            expect(response).to render_template 'shared/error'
+          end
+        end
+
+        context 'for valid access' do
+          let!(:guild_role) { create :guild_role, guild: guild, character: user_character, name: 'gm' }
+          let(:request) { delete :destroy, params: { locale: 'en', id: guild_invite.id } }
+
+          it 'deletes guild invite' do
+            expect { request }.to change { GuildInvite.count }.by(-1)
+          end
+
+          it 'and redirects to guild management page' do
+            request
+
+            expect(response).to redirect_to management_guild_en_path(guild.slug)
+          end
+        end
+      end
+    end
+
+    def do_request
+      delete :destroy, params: { locale: 'en', id: guild_invite.id }
+    end
+  end
 end